Delete .env
It turns out your code is fine, but you put .env in the wrong place, so
the `dotenv.config()` function couldn't find it.

I guess when that function fails, it doesn't throw an error, but instead
returns an object with an `error` field on it. So you have to manually
check if there was an error, otherwise it fails silently and doesn't add
anything to `process.env`.

It looks like it expects the .env file to be in the PWD. That would be
the root directory of the repo in this case.

Anyway, thanks for the api key <3